cairo: Add interactive resize support
Currently only does top/bottom/right/left because finding the corner.
Adding corner dragging is not hard, but still harder than not, so doing
that later.

This has the problem that the demo cannot set a user-data of a
wl_surface. Currently libwayland-client has no way to let a
toolkit/application to know which one (toolkit or application) owns a
surface.