libdecor: store and load floating window dimensions
After an unset_maximized event, the configure callback may be called with
width and height set to 0. This requests the client to determine the window
dimensions by itself. libdecor needs to store the current dimensions when
the surface is reconfigured and not in maximized, fullscreen or tiled mode.
These dimensions will be restored, once a configure event with no width and
height is received.