Let plugins handle dispatching
A plugin might need to add non-wl_display based dispatching, such as
timeouts or idle callbacks. For this to be possible, we need to have the
application integrate the "main loop" of libdecoration into its own. The
common way of doing this is exporting a file descriptor for all events
related to the context.

It must be the plugin that reads and dispaches events for the
wl_display, as some potential future plugins (e.g. gtk, Qt) need to
handle their own Wayland event processing.