Add API to translate to window geometry coordinates
An application may want to e.g. place a popup according relative to the
window geometry of the toplevel window; for this it needs to be able to
translate a coordinate into the window geometry coordinate space, from
the content / surface coordinate space.

Add an API that does this. How it's implemented is specific to the
plugin, so have the plugin implement it.