Add support for application popup grab dismissing
When an application grabs a popup, it likely doesn't want the window
buttons etc to work; clicking anywhere there should dismiss the popup
rather than trigger the effect of the button. The same applies to
highlight effects.

To make this possible, add a "popup grab" that implements this
behaviour. When grabbed, the plugin will act as nothing is "highlighted"
(no widget has focus), and any button click event anywhere results in
the calling of libdecor_frame_dismiss_popup(), which will invoke the
corresponding vfunc in the frame interface.