libdecor: provide window size from content and state
When the library needs to know the dimensions of the toplevel surface, e.g.
for setting minimum and maximum toplevel limits, it has to convert content
dimensions into window dimensions. Only the plugins know about their window
dimensions so we have to provide a new interface to ask a plugin about the
window dimensions for a given content dimension and window state.