cairo: reset pointer focus if focused frame is destroyed
If a frame gets destroyed while one of its wayland surfaces is focused via
the pointer, the user data of this surface will point to deallocated memory.
If the pointer focus does not change between the deallocation and the next
seat event of another frame, this will result in accessing the deallocated
memory.

To prevent this issue, reset the pointer focus if the currently focused
frame gets destroyed.