libdecor: constrain if the configured window state is floating
Previously we decided to constrain if 'libdecor_frame_is_floating' returns
true. But this only looks at the already applied window state of the frame.
At the time when 'libdecor_configuration_get_content_size' is called and a
'libdecor_configuration' is provided, this is the old state. So we should
really use the configured window state in the 'libdecor_configuration' to
decide if the window is going to be floating or not.