Move plugin base fields into private struct
The interface pointer itself is only used internally, no need to expose
it anywhere else, so move it to the private struct.

The private struct pointer is renamed to 'priv' to avoid potential
issues with e.g. C++. Plugins are now also changed to initialize the
base layer of the plugin with a helper, that allocates the private
structure instance and sets the iface pointer.