build: Make library distinguished by major version number
To do this, add the major version in the file name, and apply libtool
semantics to the shared object versions. The soversion is hard coded to
0 and will remain so, due to the name itself having the major version.

This means that e.g. libdecor-0.1.0 produces libdecor-0.so.0.100.0, a
bug fix version libdecor-0.1.1 would produce libdecor-0.so.0.100.1.

A new ABI compatible feature release 0.2.0 would produce
libdecor-0.so.0.200.0.

A new ABI incompatible major release 1.0.0 would produce
libdecor-1.so.0.0.0, 1.1.0 would produce libdecor-1.so.0.100.0, and so
on.