libdecor: use border sizes to compute all content and window conversions
The plugin interface provides functions to convert between content and
window dimensions. These functions are to a large part redundant as they
take either of the dimensions and infer the other from the border size.
This simplifies the interface by replacing the redundant plugin conversion
functions by a single interface to query the border sizes, and converts
the dimensions inside the library using either size and the plugin-specific
border size.