cairo: Keep a ref to the frame while invoking libdecor_frame_close().
The problem to be taken care of is that libdecor_frame_close()
may or may not close the decorated window and thus destroy its frame.
Subsequent code should then not access the frame it it was destroyed.

The solution is to proceed as follows
libdecor_frame_ref()   <--- the frame's reference count goes to 2
libdecor_frame_close() <--- the frame's reference count goes to 1
update the frame       <--- the frame has not been destroyed yet, so it can be updated
libdecor_frame_unref() <--- the frame's reference count goes to 0, and the frame is destroyed.

This solution is also compatible with cases where libdecor_frame_close() triggers a
"Confirm this window is to be closed" which ends up keeping the window alive.