libdecor: Exit plugins elegantly if we don't have required globals
We can't use libdecor_notify_plugin_error as context->plugin is nullptr and also we don't want it to be fatal.

Simply fall back to the next plugin if we can.

In gamescope's case, we don't have the globals that cairo/gtk backends want, so we want to fallback to dummy.